Entities, backslash
More work to process inline markup, starting with entities and escapes.
Both are written with an attempt to be commonmark spec compliant.

In addition, the inline mechanism looks like it might be a good
mechanism to handle \r line ends and carriage returns (as opposed to
the preprocessing pass that hoedown uses).