Emphasis and strong emphasis
A reasonable but not fully spec-compliant implementation of emphasis
and strong emphasis. The left and right flanking predicates are
implemented, but complex nesting behavior is not supported (there's
more state than just a scan for matching close; probably need a
stack). Also, only ASCII characters are supported for whitespace and
punctuation tests, not Unicode.

176/545 tests passing