Nesting behavior for emphasis
Algorithm for correct nesting didn't turn out to be too hard, a stack
did the trick. When a delimeter run can close, pop as much as possible
from the stack. If the stack empties, tag is closed, using strong if
more than one can be popped. When a delimeter can open but not close,
push onto the stack.

211/545 tests passing