Redo block end
Determination of block ends is now done from processing a newline,
rather than from scanning the entire block at the beginning. This will
make it easier to support different block types, and nested blocks,
and should also be faster.

Also added a SoftBreak/HardBreak distinction, to be closer to the
cmark AST.

This shouldn't be a visible change, so still 211/545 tests passing