Links and images
Added links, and a number of other changes. Events and tags now use
Cow<str> rather than &str, as unescaping entities and backslashes
results in strings other than a slice of the input. Also we now
resolve entities and character references, rather than passing them
through to HTML.

Images are parsed, very similar to links, but markup in alt text is
not suppressed. In addition, the rule against nesting in links is not
enforced.

319/545 tests passed