Lists mostly working
Adding lists, but some work remaining. In this snapshot, all lists are
considered loose. We do handle ordered and unordered lists, enforce
that all items in a list must have the same delimiter, and also end a
list with two blank lines.

Tight lists will require two passes, which is a nontrivial change.

321/545 tests passed