Loose and tight lists
Detect which lists are loose (by presence of blank line), store in
parse info, and apply in second pass (suppressing paragraph tags).

A number of other tweaks, increasing the number of tests passing.
These include fresh lines for block tags in html rendering, and lots
of fixes to handling of blank lines and indents.

387/545 tests passed