Link references
Support for collecting link reference definitions in the first pass
and resolving link references in the second. Mostly to spec, but some
gaps regarding nesting of links.

Also, we now do two full passes (rather than just collecting events),
because link references are needed for deciding how to parse links.
This is a bit slow, and there is potential for optimization: mostly,
focus on blocks in the first pass (collecting references and
identifying loose lists), and inline only in the second. Further,
suppression of <p> tags could be moved into the RawParser instead of
post-processed.

467/548 tests passed