Update to version 0.20 of CommonMark spec
Version 0.20 of CommonMark contained a tweak to lists with indent,
and this patch tracks the change.

Also, limiting bracket nesting in lists to some reasonable number,
preventing pathologically slow behavior at high nesting levels (the
algorithm is O(n^3). Nesting of inline markup is still an issue.

Also bumping to 0.0.3 and pushing to cargo.

553/553 tests passed