Initial commit of fancy-regex module
Starting implementation of a hybrid regex engine that supports both a
richer regex language and the highly efficient matching of a finite
automaton approach. See the README.md for more explanation.

This is a development snapshot, missing many features required for
production, but it still supports a fairly large fragment of the regex
language.

Ultimately this module will support static syntax highlighting of
code blocks. A common format for syntax highlighting grammars is that
of TextMate, which has been adopted by many other editors.