Standardize use of newlines in tests
The tests generated by build.rs included a newline at the end of each
HTML output string, but not at the end of each markdown input string.
Resolving this difference fixes a significant number of test cases that
are curently failing.

Before this change, 506/621 tests from spec.rs were passing when run
by `cargo test` or `pulldown-cmark -s`, but 566/621 tests were passing
using CommonMark's `spec_tests.py` (note that both were run using the
current spec.txt from the pulldown-cmark repository, since the spec.txt
in jgm/CommonMark:master has some changes since v0.27).  Github user
kivikakk surmised in #57 that the discrepancies were due to
normalization issues, but a glance at the test outputs showed that a
great number of the failing tests were only missing a trailing newline.

The only code changes here are those in build.rs and main.rs.  The other
files with changes are automatically generated by build.rs.  They do
look a little odd with extra newlines in every test string, but since
the files are autogenerated, this isn't a big deal.

The only change in build.rs is to keep the last newline of the markdown
input string attached rather than discarding it.  The same change was
made in main.rs's `run_spec` for tests run with `pulldown-cmark -s`.

The result of this change is that both testing procedures now pass
564/621 tests. This is much closer to the 566/621 from the CommonMark
`spec_tests.py` tester.

Trivial differences in HTML output aren't really what the tests are
looking for, so modifying the testing procedure seems more appropriate
than trying to change the output of the library.

There are two solutions that come to mind: trimming trailing whitespace
when comparing the output to the expected, and adding an extra newline
at the end of the input strings.  This commit only does the latter, but
the former may be desirable as well in the future.

Adding trailing newlines to all the test case input strings generated by
build.rs immediately brings the number of passing tests to 564/621.

In addition to passing 58 more tests, this brings a number of the still
failing tests closer to passing.  For instance, Example 1 before this
change ended in `bim</code></pre>`.  After this change it ends in
`bim\n</code></pre>`.  The extra newline at the end of the input string
results in a newline between `bim` and `</code>` in this test case,
which matches spec.txt's behavior.

Note that Example 140 and Example 288 pass in the CommonMark
`spec_tests.py` tester but not in `cargo test`.  At least in the case of
Example 140, this does seem to be likely due to HTML normalization
issues as mentioned by kivikakk in #57.  Whether or not a full
normalization routine is necessary is still not clear to me.