Handle Indented Code Blocks
Each line in the code block is added as a text node child of the ICB
parent node. Trailing blank lines are detached once the end of the ICB
is found.

CommonMark spec requires that we synthesize a newline at the end of a
code block that is not already terminated.

Some changes to the test harness around handling of whitespace are
required to make this pass additional tests, but there are no
regressions.