Scan for expected container marks and new containers
Starting to handle containers. On new lines we check that all expected
container marks are present, and close containers if they aren’t. Then
we iteratively scan for the start of new container blocks and add them
to the tree. Then we move on to scan for leaf blocks.

Paragraphs remove container marks if found and lazily continue the
containers if not.

Only blockquote containers are being created so far.

Passing 286/621 spec tests.