Convert fenced code block handling to line-by-line
I’m trying to convert most parsing to single line to simplify container
handling. Fenced code block beginnings and endings are now detected the
same way as other containers. We can inspect the parent node to see if
the line being processed is a continuation of an open code block.