Add non-html generating HtmlBlock tag
I want the tree to have a node type that parents html and records the
end tag required to close html blocks. This HtmlBlock item does not
itself generate any HTML in the parser output. But it needs to map to
a tag during tree parsing. I’ve added that tag to the HTML generator to
complete the match blocks, but it does not generate any actual html.

This is the first change I’ve made to html.rs. I’m pretty confident that
it does not break anything, but it probably warrants careful review
because it is client-facing.

Still passing 356/621 spec tests.