Move code fence end detection out of scan_containers
At the end of a code fence you need to move to the start of the next
line, rather than return to the beginning of the current line. That’s
very different from how the end of the other containers is handled, so
it doesn’t really work to deal with it in scan_container.

Now passing 360/621 spec tests.