Added the TeX-like math syntax extension.
This syntax is modelled after Pandoc's math support.

Both `$inline$` and `$$display$$` forms are supported and (by default)
rendered to `<span class="math math-inline">inline</span>` and `<span
class="math math-display">display</span>` respectively. Custom renderers
can be used to pre-render them as images or use different HTML classes.

For inline math the opening dollar should not be followed by whitespaces
and the closing dollar should not be followed by digits (to make `$12
and $34` not math); display math has no such restrictions. In addition,
display math can have embedded empty lines if the parent container can
span multiple lines.

This syntax is notably different from Pandoc's in that *non-escaped
dollars in the inline math are always bad*. Pandoc tries to understand
complicated TeX tokens like `$\text{the number of digits in $x$}$` but
this implementation doesn't, so the example will render as two slices of
inline math. It is recommended to use `\(\)` in dollars instead;
unfortunately the support for `\(\)` in the JS-based TeX preprocessors
is not yet widespread (at the time of writing, MathJax supports it while
KaTeX doesn't).