Remove performance warning for codeblocks
For pathological inputs for the form "` `` ``` ```` ````` ..", the
parse time should stil be linear in the length of the input.

Note that for a sequence of n codeblocks of increasing length, their
total length is n * (n + 1) / 2 + n. While we scan until the end of
file when we encounter a codeblock, we do this on the AST, and not
on the input. This means that this is quadratic in the number of
codeblock opening elements and hence linear in the size of the input.