change the special bytes LUT to a static
It was being recreated on the stack each time parse_line was called.

This will make it faster when more special chars are added, and to swap out LUTs for different parse options without affecting the performance of the first pass when those options are not enabled.

See https://github.com/raphlinus/pulldown-cmark/issues/119#issuecomment-552158266