Implement switching special-character LUTs depending on options passed
If options == Options::empty(), then the iterator does not emit ~ or |.

The iterators are still inlined all the way down, so there is no
overhead from passing the LUT pointers around, only from storing it on
the FirstPass struct.

This is slightly slower on crdt_parse (empty Options) than the previous
commit, in exchange for being more resilient in the face of extra
extensions being added. We're not fiddling with big amounts here.