Fix bug with tab indented codeblocks being interpreted as lists
This was fairly subtle. We would scan up to 3 spaces, consuming
a tab and leaving one space in the buffer. But it would also
already bump the internal text pointer forward past the tab, so
if we'd find a list marker there, it would incorrectly be
interpreted as the list because we didn't look at the space buffer.