Bump MSRV to 1.42
This allows us to use the cargo lock file V2 and the `matches!` macro,
avoiding issues with forward compatibility and clippy warnings.

Version 1.42 is a jump up from 1.36, but it is still 6 versions behind
the most recent stable version of Rust. This allows for plenty time
for pulldown users to upgrade.