Allow configuring HTML writer for footnotes
Ultimately, it may be possible to work around this, but at present I'm
not clear on that, especially without *other* breaking changes. This
preserves the existing public API and simply adds handling for this
scenario.

One option, which *would* be a breaking change but would also address a
long-standing request (#142) is to simply add support for back-links in
*all* footnote types. In that case, there would once again be no need
for configuration. However, that may need to be an *option*, since it's
otherwise a breaking change for all existing users (including any who
may be adding those manually!).