Introduce ParserBuilder and broken footnote callbacks
Add one callback each for:

- footnotes references without a corresponding definition
- footnote definitions which are entirely unreferenced

For both, introduce corresponding types (which will also be useful
later when actually doing the tree rewriting).

Given the additional complexity this requires, instead of adding yet
more `new_with_*` variations, introduce a `ParserBuilder` which allows
a user to set up a `Parser` instance like so:

    let parser = ParserBuilder(text)
        .with_options(some_options).
        .with_broken_link_callback(a_broken_link_callback)
        .with_broken_footnote_callback(a_broken_footnote_callback)
        .build();

(This can also work for the existing cases, but they are left as is for
this change.)

To enable making the builder robust, extract a new type alias,
`BrokenLinkCallbackFn`, representing the type which is wrapped in an
`Option` by the existing `BrokenLinkCallback` type, and do the same for
the two callbacks allowed.