Add data structures for standard footnote parsing
Emitting "standard" footnotes requires rewriting the tree by moving the
body of each definition to an appropriately-ordered list at the *end*
of the tree, rather than inline in the body of the document.
Accordingly, introduce two new data structures in the parser module:

- `ParsedFootnotes`: a `struct` the parser can push definitions into as
  it traverses the tree and emits items initially. When encountering a
  footnote definition or a footnote reference, it pushes them into the
  appropriate field on the `ParsedFootnotes` type:

    - a `HashMap` for the definitions, where the key is the definition
      name and the value its body

    - a `Vec` of references, so that the order which footnote references
      appear in the body of the text is maintained so that they can be
      numbered appropriately and

- `EmittableFootnotes`: a `struct` with three fields:

    - a `Vec` of footnote definitions used in the text, consisting of:

        - the `Vec<Item>` making up the body of the definition

        - the corresponding `Vec` of back-reference names, which can be
          used to generate back-reference links

    - a `Vec` of undefined references, which can be normalized, warned
      on, or any other option using the callback introduced in the
      previous commit

    - a `HashMap` of definition names and bodies for unused footnote
      definitions, which can likewise be

This provides all the information needed to correctly emit footnotes
definitions in the correct order and with the correct structure at the
end of the parse tree. (It does *not* include enough information to
emit the correct numbers for footnote reference text, but I *believe*
the existing parser infrastructure already handles that.)