Replace footnote definitions with empty nodes
This isn't really *preferable*, but is probably roughly necessary as a
stopgap? It's mildly annoying that as a consumer you would just emit an
empty string, but as it currently stands the parser (quite reasonably!)
has no way to say "nothing to see here"; it needs to actually move the
cursor in the tree. So maybe that needs to be a loop instead?