Check indentation on the closing fence relative to the line
This changes the interpretation of this document:

````markdown
 ```
    ```
````

Without this fix, pulldown-cmark thinks the second line has three
spaces of indentation, and is a valid closing code fence.
Most other commonmark implementations interpret it as having four,
because they start counting at the beginning of the line instead
of the opening fence's indentation level.