Extract non-generic ParserInner
The current `Parser`'s generics are a source of longer compile times due
to each instantiation needing to compile and optimize separate copies of
very large and complex functions.

There are good reasons to keep the `BrokenLinkCallback` as a generic
argument (see #701), so to provide the same API while reducing compile
times, extract the non-generic bits of the parser into a separate
struct. The new `ParserInner`'s methods take a `dyn BrokenLinkCallback`.

And while we're at it, let's unify the iterator impls for `Parser` and
`OffsetIter`, so that the logic is not duplicated.